refactor: 정상적인 경우의 인수테스트 모두 통과할 수 있도록 코드 리팩토링 (제약 사항 추가 전)

+++ src/test/java/kitchenpos/ui/menu/MenuRestControllerAcceptanceTestFixture.java
@@ -92,7 +92,7 @@ class MenuRestControllerAcceptanceTestFixture extends IntegrationTestHelper {
         assertSoftly(softly -> {
             softly.assertThat(result).hasSize(2);
             softly.assertThat(result.get(0).getMenuGroupId()).isEqualTo(menu.getMenuGroupId());
-            softly.assertThat(result.get(0).getPrice().longValue()).isEqualTo(menu.getPrice());
+            softly.assertThat(result.get(0).getPrice().longValue()).isEqualTo(menu.getPrice().longValue());
             softly.assertThat(result.get(0).getMenuProducts()).hasSize(1);
         });
     }
